Refactor UI into package
